security view clean up
