+++ my-video-room/module/security/class-security.php
@@ -14,7 +14,7 @@ use MyVideoRoomPlugin\Library\Dependencies;
 use MyVideoRoomPlugin\Module\Security\DAO\DBSetup;
 use MyVideoRoomPlugin\Module\Security\Library\PageFilters;
 use MyVideoRoomPlugin\Entity\MenuTabDisplay;
-use MyVideoRoomPlugin\Module\Security\Library\TemplateIcons;
+use MyVideoRoomPlugin\Module\Security\Library\SecurityNotifications;
 use MyVideoRoomPlugin\Module\Security\Shortcode\SecurityVideoPreference;
 
 /**
@@ -79,14 +79,21 @@ class Security {
 	 * Required for Normal Runtime.
 	 */
 	public function init() {
-		// Turn on Runtime Filters.
+		// Turn on Init Filters.
 		Factory::get_instance( PageFilters::class )->runtime_filters();
 		$this->security_menu_setup();
 		add_filter( 'myvideoroom_sitevideo_admin_page_menu', array( $this, 'render_security_sitevideo_tabs' ), 20, 2 );
+
 		// Add Permissions Menu to Main Frontend Template.
 		add_filter( 'myvideoroom_main_template_render', array( $this, 'render_shortcode_security_permissions_tab' ), 40, 4 );
+
 		// Add Permissions Icons Status to Main Shortcode Header.
-		\add_filter( 'myvideoroom_template_icon_section', array( Factory::get_instance( TemplateIcons::class ), 'add_default_video_icons_to_header' ), 10, 4 );
+		\add_filter( 'myvideoroom_template_icon_section', array( Factory::get_instance( SecurityNotifications::class ), 'add_default_video_icons_to_header' ), 10, 4 );
+
+		// Add Permissions Notification of Status to Main Permissions SecurityVideoPreference Normal and Admin Forms.
+		\add_filter( 'myvideoroom_security_admin_preference_buttons', array( Factory::get_instance( SecurityNotifications::class ), 'show_security_admin_status' ), 10, 3 );
+		\add_filter( 'myvideoroom_security_settings_preference_buttons', array( Factory::get_instance( SecurityNotifications::class ), 'show_security_settings_status' ), 10, 3 );
+		\add_filter( 'myvideoroom_security_roomhosts_preference_buttons', array( Factory::get_instance( SecurityNotifications::class ), 'show_security_roomhosts_status' ), 10, 3 );
 
 	}
