+++ my-video-room/module/security/views/shortcode-securityvideopreference.php
@@ -8,26 +8,22 @@
  * @package MyVideoRoomExtrasPlugin\Views\Public
  */
 
-use MyVideoRoomPlugin\SiteDefaults;
 use MyVideoRoomPlugin\Factory;
-use MyVideoRoomPlugin\Library\Dependencies;
+use MyVideoRoomPlugin\Library\HTML;
 use MyVideoRoomPlugin\Library\HttpPost;
 use MyVideoRoomPlugin\Module\Security\Entity\SecurityVideoPreference;
-use MyVideoRoomPlugin\Module\Security\Dao\SecurityVideoPreference as SecurityVideoPreferenceDAO;
 use MyVideoRoomPlugin\Module\Security\Settings\Field as InputField;
-use MyVideoRoomPlugin\Module\Security\Templates\SecurityButtons;
 
 return function (
 	?SecurityVideoPreference $current_user_setting,
 	string $room_name,
 	int $id_index = 0,
+	string $roles_output,
 	int $user_id = null
 	): string {
-	wp_enqueue_style( 'myvideoroom-template' );
-	wp_enqueue_style( 'myvideoroom-menutab-header' );
 	ob_start();
 
-	$html_library = Factory::get_instance( \MyVideoRoomPlugin\Library\HTML::class, array( 'security' ) );
+	$html_library = Factory::get_instance( HTML::class, array( 'security' ) );
 
 	/**
 	 * This should be moved to the controller.
@@ -51,65 +47,19 @@ return function (
 			$output = str_replace( '-', ' ', $room_name );
 			echo esc_attr( ucwords( $output ) );
 			?>
-				</h1>
-				<?php
-				// room permissions info.
-				$site_override              = Factory::get_instance( SecurityVideoPreferenceDAO::class )->read_security_settings( SiteDefaults::USER_ID_SITE_DEFAULTS, SiteDefaults::ROOM_NAME_SITE_DEFAULT, 'site_override_enabled' );
-				$room_disabled              = Factory::get_instance( SecurityVideoPreferenceDAO::class )->read_security_settings( $user_id, $room_name, 'room_disabled' );
-				$anonymous_enabled          = Factory::get_instance( SecurityVideoPreferenceDAO::class )->read_security_settings( $user_id, $room_name, 'anonymous_enabled' );
-				$allow_role_control_enabled = Factory::get_instance( SecurityVideoPreferenceDAO::class )->read_security_settings( $user_id, $room_name, 'allow_role_control_enabled' );
-				$block_role_control         = Factory::get_instance( SecurityVideoPreferenceDAO::class )->read_security_settings( $user_id, $room_name, 'block_role_control_enabled' );
-				$output                     = null;
-
-				if ( ! $site_override ) {
-					if ( ! $room_disabled ) {
-						$output .= '<p class="mvr-main-button-enabled" >' . esc_html__( 'Room Enabled', 'my-video-room' ) . '</p>';
-					} else {
-						$output .= '<p class="mvr-main-button-disabled" >' . esc_html__( 'Room Disabled', 'my-video-room' ) . '</p>';
-					}
-					if ( Factory::get_instance( Dependencies::class )->is_buddypress_active() ) {
-						$restrict_group_to_members_enabled = Factory::get_instance( SecurityVideoPreferenceDAO::class )->read_security_settings( $user_id, $room_name, 'restrict_group_to_members_enabled' );
-						$restrict_to_friends               = Factory::get_instance( SecurityVideoPreferenceDAO::class )->read_security_settings( $user_id, $room_name, 'bp_friends_setting' );
-						if ( $restrict_group_to_members_enabled ) {
-							$output .= '<p class="mvr-main-button-notice">' . esc_html__( 'Restricted to Members', 'my-video-room' ) . '</p>';
-						}
-						if ( $restrict_to_friends ) {
-							$output .= '<p class="mvr-main-button-notice">' . esc_html__( 'Restricted to Friends', 'my-video-room' ) . '</p>';
-						}
-					}
-					if ( $allow_role_control_enabled ) {
-						$db_setting = Factory::get_instance( SecurityVideoPreferenceDAO::class )->read_security_settings( $user_id, $room_name, 'allowed_roles' );
-						if ( ! $db_setting ) {
-							$db_setting = 'No One';
-						}
-						if ( $block_role_control ) {
-							$output .= '<p class="mvr-main-button-notice">' . esc_html__( 'Member Restrictions are Excluding : ', 'my-video-room' ) . $db_setting . '</p>';
-						} else {
-							$output .= '<p class="mvr-main-button-notice">' . esc_html__( 'Member Restrictions Only Allowing : ', 'my-video-room' ) . '' . $db_setting . '</p>';
-						}
-					}
-
-					if ( $anonymous_enabled ) {
-						$output .= '<p class="mvr-main-button-notice">' . esc_html__( 'Anonymous Disabled', 'my-video-room' ) . '</p>';
-					}
-				} else {
-					$output .= Factory::get_instance( SecurityButtons::class )->site_wide_enabled( 'nourl' );
-					$output .= '<p class="mvr-preferences-paragraph">' . esc_html__( 'An Administrator is overriding your settings with ones applied centrally. Certains Settings stored here may not be applied', 'my-video-room' ) . '</p>';
-				}
-				// phpcs:ignore -- WordPress.Security.EscapeOutput.OutputNotEscaped - Not Needed already escaped.
-				echo '<div class="mvr-button-table"> ' . $output . ' </div>';
-				?>
+			</h1>
+			<?php
+			$output = null;
+			$output = apply_filters( 'myvideoroom_security_settings_preference_buttons', $output, $user_id, $room_name );
+			// phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- function escaped upstream.
+			echo '<div class="mvr-button-table"> ' . $output . ' </div>';
+			?>
 				<form method="post" action="">
 				<input name="myvideoroom_security_room_name" type="hidden" value="<?php echo esc_attr( $room_name ); ?>" />
 				<input name="myvideoroom_security_user_id" type="hidden" value="
 					<?php
-					if ( Factory::get_instance( Dependencies::class )->is_buddypress_active() ) {
-						if ( function_exists( 'bp_is_groups_component' ) && bp_is_groups_component() ) {
-							global $bp;
-							$group_id = $bp->groups->current_group->id;
-							echo esc_attr( $group_id );
-						}
-					}
+						$user_id = apply_filters( 'myvideoroom_security_admin_preference_user_id_intercept', $user_id );
+						echo esc_html( $user_id );
 					?>
 					"/>
 					<hr>
@@ -177,7 +127,6 @@ return function (
 						name="myvideoroom_security_allowed_roles_preference[]"
 						id="myvideoroom_security_allowed_roles_preference">
 					<?php
-					$roles_output = Factory::get_instance( SecurityVideoPreferenceDAO::class )->read_multi_checkbox_admin_roles( $user_id, $room_name, 'allowed_roles' );
 					//phpcs:ignore --WordPress.Security.EscapeOutput.OutputNotEscaped - Function already escapes HTML properly upstream.
 					echo $roles_output;
 					?>
